requirements.txt
langchain_google_genai
langchain
streamlit
PyPDF2

-e .